refactor: remove facade re-exports from auto.ts — import directly from source modules
